requirements.txt
PyYAML
yconf
pymongo
elasticsearch-dsl